UP-2823 - #54 - Fixed issue that prevents creating new queues, exchangers and/or bindings.
It was being overwritten, as we were using its reference instead of a new instance.
In addition, the bean name was being set after calling afterPropertiesSet, which caused the registered bean to have no name.
Those two problems cause that during the declaration of the queues,exchange or bindings, the registered RabbitAdmin was not identified as DeclarableAdmins.
RabbitAdmin filter the declarebles here: RabbitAdmin.filterDeclarables(.....) and the admin checks here: declarableByMe(...)

Introducing Github Actions to build, release and publishing artifacts to maven central